Move code for handling exceptions to Controller.java
